impl(internal/language): move NewTestAPI
Now that golang and rust are in the same package, move
genclient.NewTestAPI to package language and unexport it, since that is
the open place this function is used.
