+++ generator/internal/genclient/genclient.go
@@ -204,51 +204,3 @@ func Generate(req *GenerateRequest) error {
 
 	return nil
 }
-
-// Creates a populated API state from lists of messages, enums, and services.
-func NewTestAPI(messages []*Message, enums []*Enum, services []*Service) *API {
-	state := &APIState{
-		MessageByID: make(map[string]*Message),
-		EnumByID:    make(map[string]*Enum),
-		ServiceByID: make(map[string]*Service),
-	}
-	for _, m := range messages {
-		state.MessageByID[m.ID] = m
-	}
-	for _, e := range enums {
-		state.EnumByID[e.ID] = e
-	}
-	for _, s := range services {
-		state.ServiceByID[s.ID] = s
-	}
-	for _, m := range messages {
-		parentID := parentName(m.ID)
-		parent := state.MessageByID[parentID]
-		if parent != nil {
-			m.Parent = parent
-			parent.Messages = append(parent.Messages, m)
-		}
-	}
-	for _, e := range enums {
-		parent := state.MessageByID[parentName(e.ID)]
-		if parent != nil {
-			e.Parent = parent
-			parent.Enums = append(parent.Enums, e)
-		}
-	}
-
-	return &API{
-		Name:     "Test",
-		Messages: messages,
-		Enums:    enums,
-		Services: services,
-		State:    state,
-	}
-}
-
-func parentName(id string) string {
-	if lastIndex := strings.LastIndex(id, "."); lastIndex != -1 {
-		return id[:lastIndex]
-	}
-	return "."
-}
